refactor: 파일명 오탈자 수정 (globalStyles -> globalStyle)
