+++ .storybook/preview.tsx
@@ -2,7 +2,7 @@ import { ThemeProvider } from "@emotion/react";
 import type { Preview } from "@storybook/react-vite";
 import React from "react";
 import theme from "../src/styles/theme";
-import GlobalStyles from "../src/styles/globalStyles";
+import GlobalStyles from "../src/styles/globalStyle";
 
 export const decorators = [
   (Story) => (
