+++ src/styles/global.ts
@@ -8,8 +8,8 @@ export const global = css`
   }
   body {
     font-family: 'Pretendard', 'Noto Sans KR', 'Apple SD Gothic Neo', sans-serif;
-    background-color: #f5f5f5;
     word-break: break-all;
+    margin: 0 auto;
   }
   a {
     text-decoration: none;
